requirements.txt
# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# AI/ML Models
dashscope>=1.20.0
pywebview>=5.0.0
volcengine-python-sdk[ark]>=1.0.0  # Volcengine Ark (豆包 SeeDance I2V); imports as volcenginesdkarkruntime

# Data Validation & Processing
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
PyJWT>=2.8.0

# Cloud Services
alibabacloud-videoenhan20200320>=1.0.0
alibabacloud-tea-openapi>=0.3.0
alibabacloud-tea-util>=0.3.0
boto3>=1.28.0

# Optional Dependencies
# Uncomment if needed:
# numpy>=1.24.0
# pillow>=10.0.0
openai>=1.0.0  # Required for DashScope OpenAI-compatible API calls
